scripts: Merge window more correctly into the group
The previous method only considered the x and y coordinates of the group and the current window resulting in it potentially moving a window into a group on another monitor.

This change incorporate the window size such that it can more correctly determine the position of a window in relation to the group.

Signed-off-by: Darrel Griët <dgriet@gmail.com>